requirements.txt
discord.py==2.6.1
flask==2.3.3
